requirements.txt
Flask
tensorflow
Pillow
requests==2.29.0
beautifulsoup4